complete crud operation and authentication with jwt token

+++ controllers/contactController.js
@@ -4,16 +4,16 @@ const { Error } = require("mongoose");
 
 // @desc GET All Contacts
 // @desc GET api/contacts
-// @access public
+// @access private
 const getContact = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
   // res.send("This is your contacts GET API");
   res.status(200).json(contacts);
 });
 
 // @desc POST All Contacts
 // @desc POST api/contacts
-// @access public
+// @access private
 const createContact = asyncHandler(async (req, res) => {
   const { name, email, phone } = req.body;
   if (!name || !email || !phone) {
@@ -25,6 +25,7 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
   // res.status(200).json(contact);
   res.status(200).json({ message: "Contact has been created", contact });
@@ -32,7 +33,7 @@ const createContact = asyncHandler(async (req, res) => {
 
 // @desc GET Single Contact
 // @desc GET api/contacts
-// @access public
+// @access private
 const getSingleContact = asyncHandler(async (req, res) => {
   console.log("Params", req.params.id);
   const contact = await Contact.findById(req.params.id);
@@ -48,7 +49,7 @@ const getSingleContact = asyncHandler(async (req, res) => {
 
 // @desc PUT single Contact
 // @desc PUT api/contacts
-// @access public
+// @access private
 
 const updateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
@@ -56,6 +57,11 @@ const updateContact = asyncHandler(async (req, res) => {
   if (!contact) {
     res.status(404);
     throw new Error("Contact Not Found");
+  }
+
+  if (contact.user_id.toString() != req.user.id) {
+    res.status(403);
+    throw new Error("You dont't have permission to update this contact");
   }
 
   const updatedContact = await Contact.findByIdAndUpdate(
@@ -69,7 +75,7 @@ const updateContact = asyncHandler(async (req, res) => {
 
 // @desc DEL single Contact
 // @desc DEL api/contacts
-// @access public
+// @access private
 
 const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
@@ -78,6 +84,10 @@ const deleteContact = asyncHandler(async (req, res) => {
   if (!contact) {
     res.status(404);
     throw new Error("Contact Not Found");
+  }
+  if (contact.user_id.toString() != req.user.id) {
+    res.status(403);
+    throw new Error("You dont't have permission to delete this contact");
   }
   const response = await Contact.deleteOne({ _id: req.params.id });
   res.status(200).json({ message: "Contact Deleted", response });
